pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Tai Lucas <tglucas@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
app = "app.__main__:main"

[tool.poetry.dependencies]
python = "^3.10"
pyzmq = "^25.0.0"
sentry-sdk = "^1.15.0"


[tool.poetry.group.dev.dependencies]
pylib = {path = "pylib"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
